fixed chats and messages

+++ app/api/chats/[id]/route.js
@@ -2,6 +2,10 @@ import Chat from '~/models/chat';
 import { connectToDB } from '~/utils/database';
 import User from "~/models/user";
 import {Schema} from "mongoose";
+import Like from "~/models/like";
+import Comment from "~/models/comment";
+import Post from "~/models/post";
+import Message from "~/models/message";
 
 export const GET = async (request, { params }) => {
   try {
@@ -58,8 +62,19 @@ export const DELETE = async (request, { params }) => {
   try {
     await connectToDB();
 
-    // Find the chat by ID and remove it
+    // Find the chat by ID
+    const chat = await Chat.findById(params.id);
+
+    if (!chat) {
+      return new Response('Chat not found', { status: 404 });
+    }
+    // Delete all messages made by the user
+    await Message.deleteMany({ chatId: params.id });
+
+    // Finally delete the user
     await Chat.findByIdAndRemove(params.id);
+
+
 
     return new Response('Chat deleted successfully', { status: 200 });
   } catch (error) {
